+++ src/main/kotlin/net/ccbluex/liquidbounce/features/module/modules/combat/crystalaura/trigger/impl/TickTrigger.kt
@@ -1,0 +1,53 @@
+/*
+ * This file is part of LiquidBounce (https://github.com/CCBlueX/LiquidBounce)
+ *
+ * Copyright (c) 2015 - 2025 CCBlueX
+ *
+ * LiquidBounce is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * LiquidBounce is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with LiquidBounce. If not, see <https://www.gnu.org/licenses/>.
+ */
+package net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.trigger.impl
+
+import net.ccbluex.liquidbounce.event.events.RotationUpdateEvent
+import net.ccbluex.liquidbounce.event.handler
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.SubmoduleCrystalDestroyer
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.SubmoduleCrystalPlacer
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.SubmoduleIdPredict
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.trigger.CrystalAuraTriggerer.offThread
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.trigger.CrystalAuraTriggerer.runDestroy
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.trigger.CrystalAuraTriggerer.runPlace
+import net.ccbluex.liquidbounce.features.module.modules.combat.crystalaura.trigger.Trigger
+
+/**
+ * Runs placing and destroying every tick.
+ */
+object TickTrigger : Trigger("Tick", true) {
+
+    @Suppress("unused")
+    private val simulatedTickHandler = handler<RotationUpdateEvent> {
+        if (offThread) {
+            runDestroy { SubmoduleCrystalDestroyer.tick() }
+            runPlace { SubmoduleCrystalPlacer.tick() }
+        } else {
+            // Make the crystal destroyer run
+            SubmoduleCrystalDestroyer.tick()
+            // Make the crystal placer run
+            SubmoduleCrystalPlacer.tick()
+            if (!SubmoduleIdPredict.enabled) {
+                // Make the crystal destroyer run
+                SubmoduleCrystalDestroyer.tick()
+            }
+        }
+    }
+
+}
